Dont allow impossible high resistances
If you press onto a big area of the screen at the same time,
sometimes the touchscreen controller gets confused and returns
erroneous x/y positions.  In this case the calculation of the
touch resistance is broken and can return impossibly high
resistances (pressure > 4200). In this case we ignore the touch
pressure as well as the x/y position.